feat(wmg): implement dataset metadata builder and refactor pipeline (#5985)
