+++ backend/wmg/data/constants.py
@@ -1,12 +1,3 @@
-# Minimum number of expressed genes for a cell to be included in the corpus.
-# See the following document for further details:
-# https://github.com/chanzuckerberg/cellxgene-documentation/blob/main/scExpression/scExpression-documentation.md#removal-of-low-coverage-cells
-GENE_EXPRESSION_COUNT_MIN_THRESHOLD = 500
-
-# Minimum value for raw expression counts that will be used to filter out computed RankIt values. Details:
-# https://github.com/chanzuckerberg/cellxgene-documentation/blob/main/scExpression/scExpression-documentation.md#removal-of-noisy-ultra-low-expression-values
-RAW_EXPR_COUNT_FILTERING_MIN_THRESHOLD = 2
-
 # wmg only includes data generated by assays that normalize for gene length
 INCLUDED_ASSAYS = {
     "EFO:0010550": "sci-RNA-seq",
