+++ backend/wmg/data/schemas/cube_schema.py
@@ -9,10 +9,12 @@ from backend.wmg.data.schemas.tiledb_filters import filters_categorical, filters
 expression_summary_indexed_dims = [
     "gene_ontology_term_id",
     "tissue_ontology_term_id",
+    "organism_ontology_term_id",
 ]
 
 expression_summary_indexed_dims_no_gene_ontology = [
     "tissue_ontology_term_id",
+    "organism_ontology_term_id",
 ]
 
 # These are the queryable cube dimensions that will be modeled as
@@ -68,6 +70,7 @@ expression_summary_schema = tiledb.ArraySchema(
 
 cell_counts_indexed_dims = [
     "tissue_ontology_term_id",
+    "organism_ontology_term_id",
 ]
 cell_counts_non_indexed_dims = expression_summary_non_indexed_dims
