+++ backend/wmg/data/transform.py
@@ -1,18 +1,14 @@
 import json
-from typing import Dict, Iterable, List, Set
+from typing import List, Set
 
 import pandas as pd
 import tiledb
 
 from backend.wmg.data.constants import CL_BASIC_OBO_NAME
-from backend.wmg.data.ontology_labels import gene_term_label, ontology_term_label
 from backend.wmg.data.snapshot import (
     CELL_TYPE_ORDERINGS_FILENAME,
     DATASET_METADATA_FILENAME,
-    EXPRESSION_SUMMARY_CUBE_NAME,
-    PRIMARY_FILTER_DIMENSIONS_FILENAME,
 )
-from backend.wmg.data.tissue_mapper import TissueMapper
 from backend.wmg.data.utils import get_datasets_from_curation_api, get_pinned_ontology_url, log_func_runtime, to_dict
 
 
@@ -209,81 +205,3 @@ def _cell_type_ordering_per_tissue(data_cells: pd.DataFrame, target_cells: List[
     tissue_cells["order"] = range(len(tissue_cells))
 
     return tissue_cells
-
-
-@log_func_runtime
-def generate_primary_filter_dimensions(snapshot_path: str, snapshot_id: int):
-    def list_primary_filter_dimension_term_ids(cube, primary_dim_name: str):
-        return cube.query(attrs=[], dims=[primary_dim_name]).df[:].groupby([primary_dim_name]).first().index.tolist()
-
-    def list_grouped_primary_filter_dimensions_term_ids(
-        cube, primary_dim_name: str, group_by_dim: str
-    ) -> Dict[str, List[str]]:
-        return (
-            cube.query(attrs=[], dims=[primary_dim_name, group_by_dim])
-            .df[:]
-            .drop_duplicates()
-            .groupby(group_by_dim)
-            .agg(list)
-            .to_dict()[primary_dim_name]
-        )
-
-    def build_gene_id_label_mapping(gene_ontology_term_ids: List[str]) -> List[dict]:
-        return [
-            {gene_ontology_term_id: gene_term_label(gene_ontology_term_id)}
-            for gene_ontology_term_id in gene_ontology_term_ids
-        ]
-
-    def build_ontology_term_id_label_mapping(ontology_term_ids: Iterable[str]) -> List[dict]:
-        return [{ontology_term_id: ontology_term_label(ontology_term_id)} for ontology_term_id in ontology_term_ids]
-
-    with tiledb.open(f"{snapshot_path}/{EXPRESSION_SUMMARY_CUBE_NAME}") as cube:
-        # gene terms are grouped by organism, and represented as a nested lists in dict, keyed by organism
-        organism_gene_ids: dict[str, List[str]] = list_grouped_primary_filter_dimensions_term_ids(
-            cube, "gene_ontology_term_id", group_by_dim="organism_ontology_term_id"
-        )
-        organism_gene_terms = {
-            organism_term_id: build_gene_id_label_mapping(gene_term_ids)
-            for organism_term_id, gene_term_ids in organism_gene_ids.items()
-        }
-
-        # tissue terms are grouped by organism, and represented as a nested lists in dict, keyed by organism
-        organism_tissue_ids: dict[str, List[str]] = list_grouped_primary_filter_dimensions_term_ids(
-            cube, "tissue_ontology_term_id", group_by_dim="organism_ontology_term_id"
-        )
-        organism_tissue_terms = {
-            organism_term_id: build_ontology_term_id_label_mapping(order_tissues(tissue_term_ids))
-            for organism_term_id, tissue_term_ids in organism_tissue_ids.items()
-        }
-
-        result = dict(
-            snapshot_id=str(snapshot_id),
-            organism_terms=build_ontology_term_id_label_mapping(
-                list_primary_filter_dimension_term_ids(cube, "organism_ontology_term_id")
-            ),
-            tissue_terms=organism_tissue_terms,
-            gene_terms=organism_gene_terms,
-        )
-
-        with open(f"{snapshot_path}/{PRIMARY_FILTER_DIMENSIONS_FILENAME}", "w") as f:
-            json.dump(result, f)
-
-
-def order_tissues(ontology_term_ids: Iterable[str]) -> Iterable[str]:
-    """
-    Order tissues based on appearance in TissueMapper.HIGH_LEVEL_TISSUES. This will maintain the priority set in
-    that class which is intended to keep most relevant tissues on top and tissues that are related to be placed
-    sequentially
-    """
-    ontology_term_ids = set(ontology_term_ids)
-    ordered_ontology_term_ids = []
-    for tissue in TissueMapper.HIGH_LEVEL_TISSUES:
-        tissue = TissueMapper.reformat_ontology_term_id(tissue, to_writable=True)
-        if tissue in ontology_term_ids:
-            ontology_term_ids.remove(tissue)
-            ordered_ontology_term_ids.append(tissue)
-
-    if ontology_term_ids:
-        ordered_ontology_term_ids = ordered_ontology_term_ids + list(ontology_term_ids)
-
-    return ordered_ontology_term_ids
