+++ backend/wmg/pipeline/utils.py
@@ -1,124 +1,124 @@
-import gc
+import json
 import logging
+import os
+import unicodedata
 
-import numpy as np
-from anndata import AnnData
-from scipy.sparse import csc_matrix, csr_matrix
+from tiledb import ArraySchema
 
-from backend.wmg.pipeline.errors import UnsupportedMatrixTypeError
+from backend.wmg.data.tissue_mapper import TissueMapper
+from backend.wmg.data.utils import (
+    create_empty_cube,
+    get_collections_from_curation_api,
+    get_datasets_from_curation_api,
+)
+from backend.wmg.pipeline.constants import (
+    DATASET_METADATA_CREATED_FLAG,
+    EXPRESSION_SUMMARY_AND_CELL_COUNTS_CUBE_CREATED_FLAG,
+    EXPRESSION_SUMMARY_DEFAULT_CUBE_CREATED_FLAG,
+    FILTER_RELATIONSHIPS_CREATED_FLAG,
+    MARKER_GENES_CUBE_CREATED_FLAG,
+    PIPELINE_STATE_FILENAME,
+    PRIMARY_FILTER_DIMENSIONS_CREATED_FLAG,
+)
 
 logger = logging.getLogger(__name__)
 
-####################################### PUBLIC FUNCTIONS #########################################
-def remap_anndata_normalized_X_to_raw_X_if_exists(adata: AnnData) -> None:
-    """
-    Remaps AnnData.X to AnnData.raw.X if and only if AnnData.raw.X exists.
 
-    Furthermore, garbage collect the original value pointed to by AnnData.X.
-
-    Parameters
-    ----------
-    data
-        The (annotated) data matrix of shape `n_obs` × `n_vars`.
-        Rows correspond to cells and columns to genes.
-    """
-    raw_expression_matrix = getattr(adata.raw, "X", None)
-
-    if raw_expression_matrix is not None:
-        logger.info("Remapping AnnData.X to AnnData.raw.X")
-        adata.X = adata.raw.X
-        gc.collect()
-
-
-def decorate_anndata_with_assays_to_filter_out(adata: AnnData, assays_to_keep: list) -> None:
-    """
-    Decorate the anndata object with a 'filter_cells' attribute that indicates which
-    rows (ie cells) to filter because the assays used are not in the given list.
-
-    Parameters
-    ----------
-    data
-        The (annotated) data matrix of shape `n_obs` × `n_vars`.
-        Rows correspond to cells and columns to genes.
-    assays_to_keep
-        The assays that are NOT in this list will be marked to be filtered out
-    """
-    assays_to_filter_out = np.logical_not(adata.obs["assay_ontology_term_id"].isin(assays_to_keep))
-
-    # Do a LOGICAL_OR to update the 'filter_cells' boolean index
-    adata.obs["filter_cells"] |= assays_to_filter_out
-
-
-def decorate_anndata_with_lowly_covered_cells_to_filter_out(adata: AnnData, min_genes: int) -> None:
-    """
-    Decorate the anndata object with a 'filter_cells' attribute that indicates which
-    rows (ie cells) to filter because they don't have a minimum number of expressed genes.
-
-    Parameters
-    ----------
-    data
-        The (annotated) data matrix of shape `n_obs` × `n_vars`.
-        Rows correspond to cells and columns to genes.
-    min_genes
-        The minimum number of expressed genes threshold for a given cell to pass
-        the filtering criteria
-
-    """
-
-    # memory efficient implementation for getting count of non-zeros per row.
-    # For CSR we use the `indptr` structure to compute number of non-zeros per row
-    # For CSC we use the `indices` structure to compute number of non-zeros per row
-    # And finally for ndarray, we simply filter the matrix and sum the resulting boolean matrix.
-    if isinstance(adata.X, csr_matrix):
-        logger.info(f"Filtering cells for anndata.X csr_matrix of shape: {adata.X.shape}")
-        gene_counts_per_cell = np.diff(adata.X.indptr)
-    elif isinstance(adata.X, csc_matrix):
-        cell_indices, counts = np.unique(adata.X.indices, return_counts=True)
-
-        # We take care of special case that an entire row has no values (implicit zero)
-        # and therefore omitted in the unique count of `adata.X.indices` by allocating
-        # a new array of size equal to the number of rows in the sparse matrix where each
-        # index in the row corresponds to a row index in the sparse matrix and then filling
-        # this new array with counts non-zero columns for a corresponding row index
-        #
-        # Example: 4X4 sparse matrix stored in column orientation where ROW 3 has no values.
-        # That is, ROW 3 is not stored in the `adata.X.indices` array:
-        # column_oriented_sparse_matrix = np.array([[1, 0, 0, 2], [0, 4, 1, 0], [0, 0, 0, 0], [0, 0, 5, 0]])
-        logger.info(f"Filtering cells for anndata.X csc_matrix of shape: {adata.X.shape}")
-        gene_counts_per_cell = np.zeros(adata.shape[0], dtype="int")
-        gene_counts_per_cell[cell_indices] = counts
-    elif isinstance(adata.X, np.ndarray):
-        logger.info(f"Filtering cells for anndata.X ndarray of shape: {adata.X.shape}")
-        gene_counts_per_cell = (adata.X > 0).sum(axis=1).flatten()
+def load_pipeline_state(corpus_path: str):
+    state_file = os.path.join(corpus_path, PIPELINE_STATE_FILENAME)
+    if os.path.exists(state_file):
+        with open(state_file, "r") as f:
+            state = json.load(f)
     else:
-        raise UnsupportedMatrixTypeError(f"Unsupported AnnData.X matrix type: {type(adata.X)}")
+        state = {
+            EXPRESSION_SUMMARY_AND_CELL_COUNTS_CUBE_CREATED_FLAG: False,
+            EXPRESSION_SUMMARY_DEFAULT_CUBE_CREATED_FLAG: False,
+            MARKER_GENES_CUBE_CREATED_FLAG: False,
+            FILTER_RELATIONSHIPS_CREATED_FLAG: False,
+            PRIMARY_FILTER_DIMENSIONS_CREATED_FLAG: False,
+            DATASET_METADATA_CREATED_FLAG: False,
+        }
+    return state
 
-    logger.info(
-        f"Shape of gene_counts_per_cell array holding counts of non-zero column values: {gene_counts_per_cell.shape}"
+
+def write_pipeline_state(pipeline_state: dict, corpus_path: str):
+    with open(os.path.join(corpus_path, PIPELINE_STATE_FILENAME), "w") as f:
+        json.dump(pipeline_state, f)
+
+
+def remove_accents(input_str):
+    nfkd_form = unicodedata.normalize("NFKD", input_str)
+    return "".join([c for c in nfkd_form if not unicodedata.combining(c)])
+
+
+def return_dataset_dict_w_publications():
+    datasets = get_datasets_from_curation_api()
+    collections = get_collections_from_curation_api()
+    collections_dict = {collection["collection_id"]: collection for collection in collections}
+
+    # cchoi: creating a helper function to format citations properly
+    def create_formatted_citation(collection):
+        publisher_metadata = collection["publisher_metadata"]
+        if publisher_metadata is None:
+            return "No Publication"
+        first_author = collection["publisher_metadata"]["authors"][0]
+        # first_author could be either 'family' or 'name'
+        citation = f"{first_author['family'] if 'family' in first_author else first_author['name']} et al. {collection['publisher_metadata']['journal']} {collection['publisher_metadata']['published_year']}"
+        formatted_citation = "No Publication" if collection["publisher_metadata"]["is_preprint"] else citation
+        return formatted_citation
+
+    dataset_dict = {}
+    for dataset in datasets:
+        dataset_id = dataset["dataset_id"]
+        collection = collections_dict[dataset["collection_id"]]
+        dataset_dict[dataset_id] = create_formatted_citation(collection)
+
+    return dataset_dict
+
+
+def list_grouped_primary_filter_dimensions_term_ids(
+    df, primary_dim_name: str, group_by_dim: str
+) -> dict[str, list[str]]:
+    """
+    This function takes a dataframe and two dimension names as input. It groups the dataframe by the second dimension name,
+    and lists the unique values of the first dimension name for each group. The output is a dictionary where the keys are the
+    unique values of the second dimension name, and the values are lists of unique values of the first dimension name for each group.
+
+    :param df: The input dataframe.
+    :param primary_dim_name: The name of the first dimension.
+    :param group_by_dim: The name of the second dimension to group by.
+    :return: A dictionary of lists of unique values of the first dimension for each group.
+    """
+    return (
+        df[[primary_dim_name, group_by_dim]]
+        .drop_duplicates()
+        .groupby(group_by_dim)
+        .agg(list)
+        .to_dict()[primary_dim_name]
     )
 
-    # boolean index encoding the cells (rows) with number of expressed genes less than the threshold
-    # and that should be marked to be filtered out
-    lowly_covered_cells_to_filter_out = gene_counts_per_cell < min_genes
 
-    adata.obs["filter_cells"] = lowly_covered_cells_to_filter_out
+def order_tissues(ontology_term_ids: list[str]) -> list[str]:
+    """
+    Order tissues based on appearance in TissueMapper.HIGH_LEVEL_TISSUES. This will maintain the priority set in
+    that class which is intended to keep most relevant tissues on top and tissues that are related to be placed
+    sequentially
+    """
+    ontology_term_ids = set(ontology_term_ids)
+    ordered_ontology_term_ids = []
+    for tissue in TissueMapper.HIGH_LEVEL_TISSUES:
+        tissue = TissueMapper.reformat_ontology_term_id(tissue, to_writable=True)
+        if tissue in ontology_term_ids:
+            ontology_term_ids.remove(tissue)
+            ordered_ontology_term_ids.append(tissue)
 
-    # Do inplace zeroing of rows that should be filtered out
-    # This is done to make the anndata object sparser and memory efficient
-    if lowly_covered_cells_to_filter_out.sum() > 0:
-        logger.info("Performing in place zeroing of AnnData.X rows to be filtered out")
-        rows_to_zero = np.where(lowly_covered_cells_to_filter_out)[0]
-        if isinstance(adata.X, csr_matrix):
-            for row in rows_to_zero:
-                adata.X.data[adata.X.indptr[row] : adata.X.indptr[row + 1]] = 0
-            adata.X.eliminate_zeros()
-        elif isinstance(adata.X, csc_matrix):
-            adata.X.data[np.in1d(adata.X.indices, rows_to_zero)] = 0
-            adata.X.eliminate_zeros()
-        elif isinstance(adata.X, np.ndarray):
-            adata.X[lowly_covered_cells_to_filter_out] = 0
-        else:
-            raise UnsupportedMatrixTypeError(f"Unsupported AnnData.X matrix type: {type(adata.X)}")
+    if ontology_term_ids:
+        ordered_ontology_term_ids = ordered_ontology_term_ids + list(ontology_term_ids)
 
-    else:
-        logger.info("Skipping inplace zeroing of AnnData.X")
+    return ordered_ontology_term_ids
+
+
+def create_empty_cube_if_needed(uri: str, schema: ArraySchema):
+    if not os.path.exists(uri):
+        logger.info(f"Creating empty cube at {uri} with schema: {schema}")
+        create_empty_cube(uri, schema)
